==> src/04_taska_finaloutputgeneration.py <==
# -*- coding: utf-8 -*-
"""04-taskA-FinalOutputGeneration.ipynb

Automatically generated by Colaboratory.
"""

import pandas as pd
import sys

classification_df = pd.read_csv(f'intermediate_outputs/{sys.argv[1]}/final_taska_classification_output_ontest.csv')
summarization_df = pd.read_csv(f'intermediate_outputs/{sys.argv[1]}/final_taska_summarization_output_ontest.csv')

new_df = pd.merge(classification_df, summarization_df, on='ID')
new_df.rename(columns = {'ID':'TestID'}, inplace = True)
new_df['SystemOutput2'] = new_df['SystemOutput2'].str.strip()

new_df.to_csv(f'final_output/{sys.argv[1]}/taskA_NewAgeHealthWarriors_{sys.argv[1]}.csv', index=False)
